requirements_supabase.txt
# NSE Supabase Integration Requirements

# Core Supabase integration
supabase>=2.0.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Optional: For advanced visualization (if needed)
# matplotlib>=3.7.0
# seaborn>=0.12.0

# HTTP requests (already included in supabase)
requests>=2.31.0
